+++ src/Sulu/Component/Webspace/Tests/Functional/Analyzer/RequestAnalyzerTest.php
@@ -214,6 +214,7 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
         $queryBag->all()->willReturn(['get' => 1]);
         $attributesBag = $this->prophesize(ParameterBag::class);
         $attributesBag->get('_sulu')->willReturn(null);
+        $attributesBag->has('_sulu')->willReturn(true);
         $attributesBag->set('_sulu', Argument::type(RequestAttributes::class))->shouldBeCalledTimes(1)->will(
             function ($arguments) use ($attributesBag) {
                 $attributesBag->get('_sulu')->willReturn($arguments[1]);
@@ -279,6 +280,7 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
         $queryBag = $this->prophesize(ParameterBag::class);
         $queryBag->all()->willReturn(['get' => 1]);
         $attributesBag = $this->prophesize(ParameterBag::class);
+        $attributesBag->has('_sulu')->willReturn(true);
         $attributesBag->get('_sulu')->willReturn(null);
         $attributesBag->set('_sulu', Argument::type(RequestAttributes::class))->shouldBeCalledTimes(1)->will(
             function ($arguments) use ($attributesBag) {
@@ -343,8 +345,10 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
         $request = $this->getMock('\Symfony\Component\HttpFoundation\Request');
         $request->request = new ParameterBag(['post' => 1]);
         $request->query = new ParameterBag(['get' => 1]);
+        $request->attributes = new ParameterBag();
 
         $this->requestAnalyzer->analyze($request);
+        $this->requestAnalyzer->validate($request);
     }
 
     /**
@@ -385,6 +389,9 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
 
         // this request will be analyzed only once
         $this->requestStack->getCurrentRequest()->willReturn($request)->shouldBeCalled();
+
+        $this->requestAnalyzer->analyze($request);
+        $this->requestAnalyzer->validate($request);
 
         $this->assertEquals('de_at', $this->requestAnalyzer->getCurrentLocalization()->getLocalization());
         $this->assertEquals('sulu', $this->requestAnalyzer->getWebspace()->getKey());